Kernel: Refactor to enable AK to compile on Aarch64
Revert changes to inline functions

Format
