+++ Kernel/Arch/x86/TrapFrame.h
@@ -9,7 +9,7 @@
 #include <AK/Types.h>
 
 #include <Kernel/Arch/x86/InterruptDisabler.h>
-#include <Kernel/Arch/x86/Processor.h>
+#include <Kernel/Arch/Processor.h>
 #include <Kernel/Arch/x86/RegisterState.h>
 
 namespace Kernel {
