+++ Kernel/Arch/x86/common/Interrupts.cpp
@@ -22,7 +22,8 @@
 #include <LibC/mallocdefs.h>
 
 #include <Kernel/Arch/x86/ISRStubs.h>
-#include <Kernel/Arch/x86/Processor.h>
+#include <Kernel/Arch/SafeMem.h>
+#include <Kernel/Arch/Processor.h>
 #include <Kernel/Arch/x86/RegisterState.h>
 #include <Kernel/Arch/x86/TrapFrame.h>
 #include <Kernel/KSyms.h>
