+++ Kernel/Arch/x86/common/Processor.cpp
@@ -19,10 +19,10 @@
 #include <Kernel/Arch/x86/CPUID.h>
 #include <Kernel/Arch/x86/Interrupts.h>
 #include <Kernel/Arch/x86/MSR.h>
-#include <Kernel/Arch/x86/Processor.h>
+#include <Kernel/Arch/Processor.h>
 #include <Kernel/Arch/x86/ProcessorInfo.h>
-#include <Kernel/Arch/x86/SafeMem.h>
-#include <Kernel/Arch/x86/ScopedCritical.h>
+#include <Kernel/Arch/SafeMem.h>
+#include <Kernel/Arch/ScopedCritical.h>
 #include <Kernel/Arch/x86/TrapFrame.h>
 
 namespace Kernel {
