+++ Kernel/Arch/x86/common/ProcessorInfo.cpp
@@ -7,7 +7,7 @@
 #include <AK/StringBuilder.h>
 #include <AK/Types.h>
 #include <Kernel/Arch/x86/CPUID.h>
-#include <Kernel/Arch/x86/Processor.h>
+#include <Kernel/Arch/Processor.h>
 #include <Kernel/Arch/x86/ProcessorInfo.h>
 
 namespace Kernel {
