+++ Kernel/Arch/x86/x86_64/Processor.cpp
@@ -5,7 +5,7 @@
  */
 
 #include <AK/StdLibExtras.h>
-#include <Kernel/Arch/x86/Processor.h>
+#include <Kernel/Arch/Processor.h>
 #include <Kernel/Arch/x86/TrapFrame.h>
 #include <Kernel/Panic.h>
 #include <Kernel/Process.h>
