+++ Kernel/Bus/PCI/Access.cpp
@@ -8,6 +8,7 @@
 #include <AK/HashTable.h>
 #include <Kernel/API/KResult.h>
 #include <Kernel/Arch/x86/IO.h>
+#include <Kernel/Arch/InterruptDisabler.h>
 #include <Kernel/Bus/PCI/Access.h>
 #include <Kernel/Debug.h>
 #include <Kernel/Firmware/ACPI/Definitions.h>
