+++ Kernel/KSyms.cpp
@@ -5,6 +5,7 @@
  */
 
 #include <AK/TemporaryChange.h>
+#include <Kernel/Arch/SafeMem.h>
 #include <Kernel/Arch/x86/SmapDisabler.h>
 #include <Kernel/FileSystem/OpenFileDescription.h>
 #include <Kernel/KSyms.h>
