+++ Kernel/StdLib.cpp
@@ -8,6 +8,7 @@
 #include <AK/MemMem.h>
 #include <AK/String.h>
 #include <AK/Types.h>
+#include <Kernel/Arch/SafeMem.h>
 #include <Kernel/Arch/x86/SmapDisabler.h>
 #include <Kernel/Heap/kmalloc.h>
 #include <Kernel/Memory/MemoryManager.h>
