+++ Kernel/Syscalls/get_stack_bounds.cpp
@@ -6,6 +6,7 @@
 
 #include <Kernel/Memory/Region.h>
 #include <Kernel/Process.h>
+#include <Kernel/Arch/x86/RegisterState.h>
 
 namespace Kernel {
