+++ Kernel/Syscalls/mmap.cpp
@@ -7,6 +7,7 @@
 
 #include <Kernel/Arch/x86/MSR.h>
 #include <Kernel/Arch/x86/SmapDisabler.h>
+#include <Kernel/Arch/SafeMem.h>
 #include <Kernel/FileSystem/OpenFileDescription.h>
 #include <Kernel/Memory/AnonymousVMObject.h>
 #include <Kernel/Memory/MemoryManager.h>
